Reverted refactor of ServerVerticle
